Added algorithm to get schools within given range
Added an algorithm "algo.py" (to be renamed) which gets a set of schools
which are laterally & longitudinally within a given range in kilometers.

Keep in mind that this script does *not* yet calculate whether a school
is in the _circular_ range.

Python modules requirements has been updated to add pyshp.